refactor: move requests pkg to internal
