+++ handler/post_handler.go
@@ -2,7 +2,7 @@ package handler
 
 import (
 	"basic_server/internal/model"
-	"basic_server/request"
+	"basic_server/internal/request"
 	"basic_server/response"
 	"basic_server/service"
 	"net/http"
