+++ server/handler/register_handler.go
@@ -4,7 +4,7 @@ import (
 	"net/http"
 
 	operror "basic_server/internal/errors"
-	"basic_server/request"
+	"basic_server/internal/request"
 	"basic_server/response"
 	"basic_server/service"
 	"basic_server/utils" //nolint
